LAB_7

+++ lab7_1.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 int main() {
     int a  =  20, b = 30;
-    double c = 4;
-    double output = a - (b/c);
+    int c = 4;
+    double output = a - (b/(double) c);
     cout << "output: " << output << endl;
     return 0;
 }
